Programmed error handling functionality, display of erros and error logging in data folder. Re-added .DS_Store file.

+++ scraper.js
@@ -2,6 +2,7 @@ const createCsvWriter = require('csv-writer').createArrayCsvWriter;
 const fs = require('fs');
 const osmosis = require("osmosis");
 
+const webAddress = 'http://shirts4mike.com/';
 const directory = 'data';
 const prices = [];
 const descriptions = [];
@@ -9,6 +10,7 @@ const imageUrls = [];
 const pageUrls = [];
 const csvData = [];
 let tempArray = [];
+let errorCount = 0;
 
 //Checks whether or not 'data' directory exists. If not, said directory is created.
 if(!fs.existsSync(directory)) {
@@ -18,7 +20,7 @@ if(!fs.existsSync(directory)) {
 //Goes to 'http://shirts4mike.com/', goes to t-shirts catalogue page, finds and returns
 //t-shirt info for each t-shirt in catalogue.
 osmosis
-	.get('http://shirts4mike.com/')
+	.get(webAddress)
 	.find('li.shirts > a')
 	.follow('@href')
 	.find('div.shirts > div.wrapper > ul > li > a')
@@ -41,28 +43,38 @@ osmosis
 		csvData.push(tempArray);
 	})
 	.done(function() {
-		if(!fs.existsSync(directory + '/' + getDate() + '.csv'))
-			createCSV();
-		else {
+		if(!fs.existsSync(directory + '/' + getDate() + '.csv') && csvData.length > 0)
+			createCSV(csvData, directory + '/' + getDate() + '.csv');
+		else if(fs.existsSync(directory + '/' + getDate() + '.csv') && csvData.length > 0) {
 			fs.unlinkSync(directory + '/' + getDate() + '.csv');
-			createCSV();
+			createCSV(csvData, directory + '/' + getDate() + '.csv');
 		}
 	})
-	// .log(console.log)
-	// .error(console.log)
-	// .debug(console.log)
-
-//Creating CSV file through use of csv-writer package and csvData array.
-function createCSV() {
-	const csvWriter = createCsvWriter({
-	    header: ['PRICE', 'TITLE', 'IMAGE URL', 'URL', 'TIME'],
-	    path: directory + '/' + getDate() + '.csv'
+	// .log(function(logInfo) {
+	// 	console.log(logInfo);
+	// })
+	.error(function(error) {
+		let errorCount = 0;
+		errorCount++;
+		logError(error, directory + '/scraper-error.log');
+		consoleLogError(error);
 	});
 
-	csvWriter.writeRecords(csvData)       // returns a promise
-	    .then(() => {
-	        console.log('...CSV File Written');
-	    });
+//Creating CSV file through use of csv-writer package and csvData array.
+function createCSV(array, filePath) {
+	try {
+		const csvWriter = createCsvWriter({
+		    header: ['PRICE', 'TITLE', 'IMAGE URL', 'URL', 'TIME'],
+		    path: filePath
+		});
+		csvWriter.writeRecords(array)       // returns a promise
+		    .then(() => {
+		        console.log('...CSV file written');
+		    });
+	} catch(error) {
+		logError(error, directory + '/scraper-error.log');
+		consoleLogError(error.toString());
+	}
 }
 
 function getDate() {
@@ -77,4 +89,23 @@ function getTime() {
 	let date = new Date();
 	let dateArray = date.toString().split(' ');
 	return dateArray[4];
+}
+
+function logError(errorMessage, path) {
+		let timeStamp = Date();
+		try {
+		  fs.appendFileSync(path, '[' + timeStamp + '] <' + errorMessage + '>\n');
+		} catch (error) {
+	  		console.error(error);
+		}
+}
+
+function consoleLogError(err) {
+	if(err.includes('(get)')) {
+		console.log('There was a 404 error. Unable to connect to ' + webAddress);
+	} else {
+		errorCount++;
+		if(errorCount < 2)
+			console.log('Whoooops! Something went wrong. Please check the "scraper-error.log" file.');
+	}
 }
